Fix the units of the Molecule geometry (#476)
* Fix the units of the Molecule geometry

PySCF's method `atom_coords()` defaults to Bohr although the rest of the
module defaults to Angstrom. This went unnoticed and is now fixed.

* Extend fix to ElectronicStructureDriverResult.from_legacy_driver_result

* Enforce Angstrom in Molecule construction

We actually need to enforce Angstrom as long as `Molecule` does not
support either units type. If we rely on `self._unit` we could end up
running into the same problem if the `PySCFDriver` was initialized with
Bohr coordinates.

* Fix copyright

Co-authored-by: Manoel Marques <Manoel.Marques@ibm.com>
Co-authored-by: Panagiotis Barkoutsos <bpa@zurich.ibm.com>