QCSchema (#765)
* feat: support basic QCSchema

Adds dataclasses to represent data stored in the QCSchema format [1].
Some basic test cases have been taken from the QCSchema test files.

For this to be usable, further changes are required extending the
QCSchema to support active-spaces, by including the two-body terms
directly in the files.
Support for HDF5 files will also need to be added.

[1]: https://molssi-qc-schema.readthedocs.io/en/latest/index.html

* HDF5 integration for QCSchema

* refactor: improve the QCSchema unittests

* test: prepare HDF5 capabilities tests

* feat: Active-Space Extension

Adds the Active-Space extension to the QCSchema as proposed here:
https://github.com/MolSSI/QCSchema/pull/83

* Add reno

* docs: add QCSchema attribute documentation

* Fix sphinx docs

* generalize type hints of QCSchema attributes

* fix: respect optionality of attributes in typehint

* refactor: split QCSchema into multiple files

The qcschema.py file has grown quite substantially. In order to improve
maintainability and class extensions down the road, this refactoring
splits that single file into multiple files.

* refactor unittest to use TemporaryDirectory

The validity of this change has been tested in #770, since the test here
is currently still being skipped.

* Add documentation on optional attributes of QCSchema

Co-authored-by: Manoel Marques <manoel.marques@ibm.com>