Refactor `BaseProblem.second_q_ops()` (#772)
* refactor: enable dict_aux_operators by default

* refactor: deprecate qiskit_nature.ListOrDict

This removes the usage of ListOrDict wherever possible.
The QubitConverter class is the last place this is needed, because the
UCC and UVCC ansatze pass their excitation operators for conversion as
a list.

That is fine for now, but may be improved when tackling #771

* refactor: match return type of `BaseProblem.second_q_ops()` to Terra

The algorithms in Terra always take a tuple of the main operator and
some auxiliary operators. This commit changes the `BaseProblem`
interface to already return the generated operators in the same format.
This simplifies the handling of the main operator, a scenario that
occurs very often and required manual intervention thus far.