Various improvements to the `QubitMapper` classes (#1044)
* Make suppress_none argument in QubitMapper API keyword-only

* Deprecate QubitMapper.allows_two_qubit_reduction

Closes #1039

* Deprecate nmodes arguments in QubitMapper classes

* feat: allow setting the operators register_length at the time of mapping

Closes #831

* docs: improve documentation of typed QubitMapper subclasses

We currently have three subclasses of the QubitMapper only used for
typing purposes. These add unnecessary noise to the documentation since
these classes cannot be used directly. Thus, removing them from the
public docs page should be fine.
This commit also re-adds the actual typing methods to these classes. In
the previous commit I removed the `_map_single` version of these which
did not add any value perceived from the end-user (because these were
private methods).

* Add release note

* docs: document inherited members in mappers module

This is more of an ad-hoc fix for the current situation. Qiskit Terra
also already differentiates in cases where they want to document the
inherited members and where they do not. However, they default to
document them and disable them as needed. For the scope of this PR, I
chose to do the inverse but we should do a follow-up and align our
documentation style with that of Qiskit Terra in the future.