Extract code example from README into docs (#1122)
* Extract code example from README into docs

We have been receiving several comments/questions/issues regarding the
code example from the README not working as expected. This is due to it
being aligned with the development and not the published code. In order
to resolve this, I am moving the code example into the documentations
"getting started" page. In this way, the README can simply link there
which ensures that the snippet is aligned with the published (stable)
code.

In order to still be able to test the proper functioning of this code, I
am using doctest and have introduced the new `formatting_precision`
attribute on the result objects.

Fixes #1099

* Update docs/getting_started.rst

Co-authored-by: Steve Wood <40241007+woodsp-ibm@users.noreply.github.com>

* docs: update optional installs sections

Closes #1123

---------

Co-authored-by: Steve Wood <40241007+woodsp-ibm@users.noreply.github.com>